Reworking the styling of the site to put most of it into css.
The CSS now limits the width of the main elements of the site on wide
screen, and is responsive to small devices.

Everything is now in CSS, except for the background images on some
buttons. Those are still inline styles due to their one-off nature.

Images have been moved to their own folder so they don't clutter up the
main HTML files.